currently string matching is working

+++ SelectionFunctions.py
@@ -2,5 +2,5 @@ from Constants import TRUNCATION_RATE
 
 
 def truncation_selection(population: list):
-    truncate_size = len(population) * TRUNCATION_RATE
+    truncate_size = int(len(population) * TRUNCATION_RATE)
     return population[:truncate_size]
